requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
channels==4.2.2
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
Django==5.1.6
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-filter==25.1
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
geographiclib==2.0
geopy==2.4.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mysqlclient==2.2.7
openai==1.76.2
packaging==24.2
pillow==11.1.0
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
